修复背包储存系统多项缺陷：原子化 removeItem 消除竞态、核心流程加事务包裹、修复灵石 long转int 截断、DB quantity 列 INT转BIGINT、效果执行顺序修正、添加批量操作 API、updated_at 自动更新、竞拍结算防并发

+++ src/main/java/com/mtxgdn/game/item/CurrencyEffect.java
@@ -42,7 +42,7 @@ public class CurrencyEffect extends ItemEffect {
             sb.append("获得了 ").append(goldGain).append(" 金币，");
         }
         if (spiritStoneGain > 0) {
-            itemService.addItem(playerId, SPIRIT_STONE_KEY, (int) spiritStoneGain);
+            itemService.addItem(playerId, SPIRIT_STONE_KEY, spiritStoneGain);
             sb.append("获得了 ").append(spiritStoneGain).append(" 灵石，");
         }
         return sb.toString();
